refactor(keyboard): カタログの宣言をlibから追い出す
lib/keyboard/global-shortcut-catalog.ts が calendar.shortcuts.actions.* という
feature の翻訳キーをリテラルで持っていた。import ではないため lint:boundaries
は素通りするが、calendar 側でキー名を変えると lib が実行時にだけ壊れる。

lib/keyboard は型・合成・キー処理の機構だけを持つ層とし、「このアプリにどの
ショートカットがあるか」は Composition Layer の content として持つ。

calendar のカタログは features/calendar/keyboard/ から lib/ へ移す。hooks/keyboard/
と同名ディレクトリが2階層に並ぶ状態を避け、標準ディレクトリ構造に揃える。
