+++ apps/product/src/app/[locale]/(app)/_overlays/app-shortcut-catalog.ts
@@ -1,6 +1,38 @@
 import { CALENDAR_SHORTCUT_CATALOG } from '@/features/calendar';
-import { GLOBAL_SHORTCUT_CATALOG } from '@/lib/keyboard/global-shortcut-catalog';
-import { mergeShortcutCatalogs } from '@/lib/keyboard/shortcut-catalog';
+import { mergeShortcutCatalogs, type ShortcutCatalog } from '@/lib/keyboard/shortcut-catalog';
+
+/**
+ * どのページでも有効なショートカットの宣言。
+ *
+ * lib/keyboard/ には置かない。あちらは型・合成・キー処理という機構だけを持つ層で、
+ * 「このアプリにどのショートカットがあるか」「そのラベルがどの翻訳 namespace に
+ * あるか」は app 側の content にあたる。lib に置くと lib が feature の翻訳キーへ
+ * 依存し、キー名の変更が実行時にしか露見しなくなる。
+ */
+const GLOBAL_SHORTCUT_CATALOG: ShortcutCatalog = {
+  groups: [
+    {
+      id: 'general',
+      labelKey: 'shortcuts.groups.general',
+      scope: 'global',
+      order: 0,
+    },
+  ],
+  entries: [
+    {
+      groupId: 'general',
+      labelKey: 'calendar.shortcuts.actions.open',
+      keys: ['Shift+?'],
+      order: 0,
+    },
+    {
+      groupId: 'general',
+      labelKey: 'calendar.shortcuts.actions.openSearch',
+      keys: ['Cmd+K'],
+      order: 10,
+    },
+  ],
+};
 
 /**
  * app全体のショートカットカタログ。
