+++ apps/product/src/features/calendar/lib/calendar-shortcut-catalog.ts
